requirements.txt

fastapi
uvicorn[standard]
pydantic
pytesseract
Pillow
spacy
dateparser
python-multipart
pytest
httpx